fix: fix load&save state

+++ mindone/diffusers/models/embeddings.py
@@ -471,7 +471,7 @@ class CogVideoXPatchEmbed(nn.Cell):
                 pos_embedding = self._get_positional_embeddings(height, width, pre_time_compression_frames)
                 pos_embedding = pos_embedding.to(dtype=embeds.dtype)
             else:
-                pos_embedding = self.pos_embedding
+                pos_embedding = self.pos_embedding.to(dtype=embeds.dtype)
 
             embeds = embeds + pos_embedding
